requirements.txt
asgiref==3.4.1
boto3==1.18.55
botocore==1.21.55
Django==3.2.7
django-filter==21.1
django-storages==1.11.1
jmespath==0.10.0
mysqlclient==2.0.3
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7